+++ packages/headless/doc-parser/mocks/mock-obj-entity.ts
@@ -1,0 +1,12 @@
+import {ObjEntity} from '../src/entity';
+
+export function buildMockObjEntity(config: Partial<ObjEntity> = {}): ObjEntity {
+  return {
+    name: '',
+    type: '',
+    desc: '',
+    isOptional: false,
+    members: [],
+    ...config,
+  };
+}