+++ packages/headless/rollup.config.js
@@ -1,12 +1,14 @@
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import typescript from '@rollup/plugin-typescript';
+import tsPlugin from '@rollup/plugin-typescript';
 import replacePlugin from '@rollup/plugin-replace';
 import {terser} from 'rollup-plugin-terser';
 import {sizeSnapshot} from 'rollup-plugin-size-snapshot';
 import alias from '@rollup/plugin-alias';
 import {resolve as pathResolve} from 'path';
+import dts from "rollup-plugin-dts";
 
+const typescript = () => tsPlugin({tsconfig: './src/tsconfig.build.json'});
 const isCI = process.env.CI === 'true';
 const isProduction = process.env.BUILD === 'production';
 
@@ -86,6 +88,16 @@ const browserConfig = {
   ],
 };
 
-const config = isProduction ? [nodeConfig, browserConfig] : [browserConfig];
+
+// Api-extractor cannot resolve import() types, so we use dts to create a file that api-extractor
+// can consume. When the api-extractor limitation is resolved, this step will not be necessary.
+// [https://github.com/microsoft/rushstack/issues/1050]
+const typeDefinitions = {
+  input: "./dist/index.d.ts",
+  output: [{file: "temp/headless.d.ts", format: "es"}],
+  plugins: [dts()]
+}
+
+const config = isProduction ? [nodeConfig, typeDefinitions, browserConfig] : [browserConfig];
 
 export default config;